requirements.txt
torch
pyyaml
higher
torchvision
tqdm
matplotlib
scikit-learn
h5py